Raw client into its own package
