+++ protocol/client/smartinfo.go
@@ -1,4 +1,4 @@
-package nanogit
+package client
 
 import (
 	"context"
@@ -38,7 +38,7 @@ import (
 //
 //	Returns an error if the HTTP request fails, the server returns a non-2xx status code,
 //	or the response body cannot be read.
-func (c *rawClient) SmartInfo(ctx context.Context, service string) ([]byte, error) {
+func (c *RawClient) SmartInfo(ctx context.Context, service string) ([]byte, error) {
 	u := c.base.JoinPath("info/refs")
 
 	query := make(url.Values)
