+++ protocol/client/uploadpack.go
@@ -1,4 +1,4 @@
-package nanogit
+package client
 
 import (
 	"bytes"
@@ -12,7 +12,7 @@ import (
 
 // UploadPack sends a POST request to the git-upload-pack endpoint.
 // This endpoint is used to fetch objects and refs from the remote repository.
-func (c *rawClient) UploadPack(ctx context.Context, data []byte) ([]byte, error) {
+func (c *RawClient) UploadPack(ctx context.Context, data []byte) ([]byte, error) {
 	body := bytes.NewReader(data)
 
 	// NOTE: This path is defined in the protocol-v2 spec as required under $GIT_URL/git-upload-pack.
